Separate out the namespace fixtures for a Bucket
